requirements/pip.txt
python-coveralls
cothread
pytest-cov
mock==2.0
